Criado diretório de controladores
> Agora, existe um diretório (Módulo) de controladores para cada entidade relevante do sistema
> Todos os possíveis problemas de imports já foram resolvidos, tanto na Main.hs quanto nos próprios controladores
