+++ Src/Controller/MonitorController.hs
@@ -1,0 +1,3 @@
+module Src.Controller.MonitorController where
+    import Src.Model.Monitor
+    import Src.Util.TxtFunctions
